Typo and grammar fixes
Not sure if English isn't your primary language, but it helps to make sure spelling and grammar issues are addressed if you use English in a readme. I initially was turned away by the "containt" typo, but then realized that probably wasn't your primary focus :)  

There are more improvements that can be made, but these are some of the easy and obvious ones.

Love the project, btw!